+++ src/components/NavigationMenu/NavigationMenu.js
@@ -13,6 +13,8 @@ import {
   getSendItemInTransitUrl,
 } from '../../constants';
 
+import styles from './NavigationMenu.css';
+
 export const getDataOptions = (intl) => ([
   {
     label: intl.formatMessage({ id: 'ui-requests-mediated.app.mediatedRequestsActivities.navigation' }),
@@ -46,23 +48,34 @@ export const handleChangeMenu = (event, location, history) => {
 
 const NavigationMenu = ({
   value,
+  separator,
 }) => {
   const intl = useIntl();
   const history = useHistory();
   const location = useLocation();
 
   return (
-    <Select
-      data-testid="navigationMenu"
-      value={value}
-      dataOptions={getDataOptions(intl)}
-      onChange={(event) => handleChangeMenu(event, location, history)}
-    />
+    <>
+      <Select
+        data-testid="navigationMenu"
+        value={value}
+        dataOptions={getDataOptions(intl)}
+        onChange={(event) => handleChangeMenu(event, location, history)}
+      />
+      { separator &&
+        <hr className={styles.separator} />
+      }
+    </>
   );
 };
 
 NavigationMenu.propTypes = {
   value: PropTypes.string.isRequired,
+  separator: PropTypes.bool,
+};
+
+NavigationMenu.defaultProps = {
+  separator: false,
 };
 
 export default NavigationMenu;