+++ src/utils.js
@@ -1,0 +1,10 @@
+export const transformRequestFilterOptions = (formatMessage, source = []) => (
+  source.map(({ label, value }) => ({
+    label: formatMessage({ id: label }),
+    value,
+  }))
+);
+
+export const getIsTitleLevelRequestsFeatureEnabled = (data) => (
+  data?.items[0]?.value?.titleLevelRequestsFeatureEnabled || false
+);