+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Add landing page actions: Confirm item arrival, Mediated requests activities and Send item in transit. Refs UIREQMED-2.
 * Hide permission that should not be visible. Refs UIREQMED-20.
 * Add caret to `react` peer dependency. Refs UIREQMED-23.
+* Add Search field for Mediated requests activity page. Refs UIREQMED-3.
+* Add Filters for Mediated requests actions page. Refs UIREQMED-4.
 
 ## 1.0.0
 * New app created with stripes-cli. Updated module after created with stripes-cli. Refs UIREQMED-1.